refactor(review-form): переносит логику компонента в отдельный хук; рефакторит проверку на количество символов
